requirements36.txt
absl-py==0.10.0
astor==0.8.1
autocorrect==2.0.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
editdistance==0.5.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
imageio==2.9.0
importlib-metadata==1.7.0
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
mongoengine==0.20.0
networkx==2.5
numpy==1.18.4
opencv-python==4.2.0.34
opt-einsum==3.3.0
Pillow==7.1.2
protobuf==3.13.0
pymongo==3.10.1
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.3
tensorflow-estimator==1.15.1
termcolor==1.1.0
tifffile==2020.8.25
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
